+++ src/arch/armv8/gicv2.c
@@ -36,7 +36,7 @@ size_t gich_num_lrs()
     return ((gich.VTR & GICH_VTR_MSK) >> GICH_VTR_OFF) + 1;
 }
 
-inline uint64_t gich_read_lr(size_t i)
+inline unsigned long gich_read_lr(size_t i)
 {
     if (i < NUM_LRS) {
         return gich.LR[i];
@@ -45,7 +45,7 @@ inline uint64_t gich_read_lr(size_t i)
     }
 }
 
-inline void gich_write_lr(size_t i, uint64_t val)
+inline void gich_write_lr(size_t i, unsigned long val)
 {
     if (i < NUM_LRS) {
         gich.LR[i] = val;
@@ -94,7 +94,7 @@ static inline void gicc_init()
 
     gich.HCR |= GICH_HCR_LRENPIE_BIT;
     
-    uint64_t sgi_targets = gicd.ITARGETSR[0] & BIT32_MASK(0, GIC_TARGET_BITS);
+    uint32_t sgi_targets = gicd.ITARGETSR[0] & BIT32_MASK(0, GIC_TARGET_BITS);
     ssize_t gic_cpu_id = 
         bitmap_find_nth((bitmap_t*)&sgi_targets, GIC_TARGET_BITS, 1, 0, true);
     if(gic_cpu_id < 0) {
@@ -230,7 +230,7 @@ void gic_set_prio(irqid_t int_id, uint8_t prio)
     gicd_set_prio(int_id, prio);
 }
 
-uint64_t gic_get_prio(irqid_t int_id)
+uint8_t gic_get_prio(irqid_t int_id)
 {
     return gicd_get_prio(int_id);
 }
